+++ Windows/QuickLookProtein.Settings/App.xaml.cs
@@ -1,0 +1,7 @@
+using System.Windows;
+
+namespace QuickLookProtein.Settings;
+
+public partial class App : Application
+{
+}